# src/SavingsAccount.Api/Models/ErrorResponse.cs
namespace SavingsAccount.Api.Models;

public class ErrorResponse
{
    public string Error { get; set; } = string.Empty;
    public string Message { get; set; } = string.Empty;
    public object? Details { get; set; }
}